+++ upload/admin/model/extension/payment/realex_remote.php
@@ -267,6 +267,6 @@ class ModelExtensionPaymentRealexRemote extends Model {
 	public function getTotalRebated($realex_order_id) {
 		$query = $this->db->query("SELECT SUM(`amount`) AS `total` FROM `" . DB_PREFIX . "realex_remote_order_transaction` WHERE `realex_remote_order_id` = '" . (int)$realex_order_id . "' AND 'rebate'");
 
-		return (double)$query->row['total'];
+		return (float)$query->row['total'];
 	}
 }
